Add HTTPClient
